puzzle size edge case
fix for when a user had previously specified a set puzzle.size but then
used one of the add, remove, replace word funcitons and the puzzle
wouldn't resize to fit the new words. For example is a user had a puzzle
of size 5 but wanted to add the word "apothecary" to the puzzle. The
word obviously wouldn't fit. By supply the reset_size variable to True,
the puzzle would be resized.